refector: 응답값에 rank 순위 추가

+++ src/main/java/project/maru/domain/Rank.java
@@ -24,7 +24,7 @@ public class Rank {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int id;
+  private Long id;
   private String eMail;
   private String userId;
   private int score;
@@ -41,13 +41,13 @@ public class Rank {
     this.score = score;
   }
 
-  public Rank(int id, String eMail, String userId, int score) {
+  public Rank(Long id, String eMail, String userId, int score) {
     this.id = id;
     this.eMail = eMail;
     this.userId = userId;
     this.score = score;
   }
 
-  public Rank(int id, String userId, int score, String eMail) {
+  public Rank(Long id, String userId, int score, String eMail) {
   }
 }
